requirements.txt
amqp==2.5.2
anyjson==0.3.3
astroid==2.3.2
Babel==2.7.0
beautifulsoup4==4.8.2
billiard==3.6.1.0
bs4==0.0.1
celery==4.4.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
Django==2.2.6
django-debug-toolbar==2.0
django-ipware==2.1.0
django-phonenumber-field==3.0.1
django-six==1.0.4
django-tracking2==0.5.0
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
joblib==0.14.0
kombu==4.6.7
lazy-object-proxy==1.4.2
lxml==4.4.2
maxminddb==1.5.2
mccabe==0.6.1
more-itertools==8.1.0
numpy==1.17.3
phonenumbers==8.11.0
Pillow==6.2.1
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytz==2019.3
redis==3.3.11
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
vine==1.3.0
wrapt==1.11.2
xlwt==1.3.0
zipp==0.6.0
